ProtocolGen: enum-typed fields and null-tolerant writes
Enum-schema'd fields are now declared AS their generated enum (writes cast
to the underlying primitive, reads cast back), so an out-of-range constant
is uncompilable instead of a silent wire value. That class of bug is real:
StartGame carried xboxLiveBroadcastSetting=6, a Pocket-era magic number
outside GamePublishSetting (0-4), and the 1.26.40 client kills the session
over it where 1.26.33 shrugged.

overrides.json gains the GameType schema bug: the schema lists Undefined
first, but engine Undefined is -1 and never on the wire; values start at
Survival=0 and include the deprecated spectator pair 3/4 the schema omits.

Writers now treat null like Cereal treats default-constructed: a null
vector writes count 0, a null struct writes as new(). McpeSetEntityData
NRE'd on every entity broadcast (nothing sets synchedProperties) and took
its whole compressed batch down with it, silently.

+++ src/MiNET/MiNET.Console/ChunkGeneratorHandler.cs
@@ -78,7 +78,7 @@ namespace MiNET.Console
 
 			Client.LevelInfo.LevelName = message.levelId;
 			Client.LevelInfo.Version = 19133;
-			Client.LevelInfo.GameType = message.settings.gameType;
+			Client.LevelInfo.GameType = (int) message.settings.gameType;
 
 			_internalStates = new HashSet<BlockStateContainer>(BlockFactory.BlockPalette);
